Backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-dotenv
mysql-connector-python